+++ .k8s/requirements.txt
@@ -1,0 +1,22 @@
+cachetools==5.2.0
+certifi==2022.9.24
+charset-normalizer==2.1.1
+click==8.1.3
+docopt==0.6.2
+google-auth==2.12.0
+idna==3.4
+Jinja2==3.1.2
+kubernetes==24.2.0
+MarkupSafe==2.1.1
+oauthlib==3.2.1
+pyasn1==0.4.8
+pyasn1-modules==0.2.8
+python-dateutil==2.8.2
+PyYAML==6.0
+requests==2.28.1
+requests-oauthlib==1.3.1
+rsa==4.9
+six==1.16.0
+urllib3==1.26.12
+websocket-client==1.4.1
+yarg==0.1.9
